[#15] Implemented percentile function
